[tserver] report the newly bootstrapped tablet after OpenTablet completes
A tablet may be deleted after starting successfully, but it couldn't be
deleted if not deleted from TransitionInProgressMap.

This patch moves one of MarkDirty calls from TabletReplica::Start,
and the report of tablet state change from BOOTSTRAPPING to RUNNING
is postponed until open tablet completes and TransitionInProgressDeleter
destroyed.

This patch also adds a regression test which would not pass without the fix.

Change-Id: I591dc8daceb81f9e800098be77c928d391cdc00a