KUDU-3346: fix rebalancer tool fails to run with '--ignored_tservers'
Prior to this patch the validity of 'ignored_tservers' was checked when
'BuildClusterinfo', which leads to a failure when the 'raw_info' only contains
contains information of tservers on a specific location. This patch fix it by
moving the parameter validity check into 'KsckResultsToClusterRawInfo', because
ksck results contain original cluster information.

I noticed 'ClusterInfo::tservers_to_empty' is not necessary to be built when
'BuildClusterInfo', because we use this info only for printing cluster's stats
and running IgnoredTserverRunner. This should be refactored in follow-up patch.

This patch adds a regression test for the issue and I also verified this fix on
a real cluster.

Change-Id: I1361f562f3e886077a79c3de8ea5fb2ebb8df6e9