KUDU-3386 Update macOS docs for ARM
This is the accompanying doc change to the ARM Mac support patch [1].
As the introductory line says, we now support ARM. Added macOS 11
(Big Sur) and macOS 12 (Monterey) to the supported macOS versions. Added
a link to the known not working tests for transparency, and to encourage
contributions. The CMake OPENSSL_ROOT_DIR flag's location in the guide
does not work on ARM Macs, as the homebrew path is different. Both
Kudu and thirdparty JWT need proper OPENSSL_ROOT_DIR. A pkg-config
export is added for OpenSSL which works on both Intel and Apple Silicon.
This takes care of both Kudu and thirdparty JWT.

[1] commit hash: 543e128d473f8f7836e605bba8cd6512fa918550

Change-Id: I02e5993979b88b525006344324e6b3be29aa869c
Reviewed-on: http://gerrit.cloudera.org:8080/18822
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>