Fix glog linker flags for ARM Mac
"-L$PREFIX/lib" in the glog build definitions was causing SIGBUS when
doing Kudu startup, on M1 Mac. The issue is reproducible with the
logging-test:
@ 0x1a09f84e4 _sigtramp
@ 0x1a0947afc std::__1::basic_ostream<>::operator<<()
@ 0x101de0fc0 kudu::operator<<()
@ 0x100b70ab4 kudu::LoggingTest_TestThrottledLogging_Test::TestBody()
@ 0x10191ceb8 testing::internal::HandleExceptionsInMethodIfSupported<>()
@ 0x10191ce00 testing::Test::Run()
@ 0x10191e460 testing::TestInfo::Run()
@ 0x10191ed78 testing::TestSuite::Run()
@ 0x10192c85c testing::internal::UnitTestImpl::RunAllTests()
@ 0x10192c0b8 testing::internal::HandleExceptionsInMethodIfSupported<>()
@ 0x10192c01c testing::UnitTest::Run()
@ 0x100e20920 RUN_ALL_TESTS()
@ 0x100e20650 main
Bus error: 10
Investigations on Linux: Removing "-L$PREFIX/lib" does not affect the
glog link process. The thirdparty libunwind.so and libgflags.so are
linked correctly as well. Only the runpath has to be specified, else ldd
shows not found for libunwind.so and libgflags.so.
By removing "-L$PREFIX/lib" from the glog build definitions Kudu works
fine on my M1 Mac.

Change-Id: I3771a086dfb7caf89e9a7fc0f328e77ee7049ec5
Reviewed-on: http://gerrit.cloudera.org:8080/18902
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>