KUDU-3571: fix flakiness in AutoIncrementingItest.BootstrapNoWalsNoData
The test AutoIncrementingItest.BootstrapNoWalsNoData sometimes failed due
to MVCC timestamp has not been initialized or inability to wait for in-flight
ops to be finished. This patch fixes this issue by waiting for everything
to be consistent before scanning.

Before this patch, when running the test with DEBUG configuration, 6/20 tests
failed. After this patch, 20/20 tests succeed.

Change-Id: I5bd387c82b632dbb77aa5a45f831273392ae05b4